+++ example/lib/main.dart
@@ -41,7 +41,7 @@ class _MyAppState extends State<MyApp> {
     _dio.interceptors.add(_alice.getDioInterceptor());
     _httpClient = HttpClient();
     _chopper = ChopperClient(
-      interceptors: _alice.getChopperInterceptor(),
+      interceptors: [_alice.getChopperInterceptor()],
     );
     _postsService = PostsService.create(_chopper);